+++ src/monetization.js
@@ -12,7 +12,7 @@ const JSONPClient = require('./JSONPClient');
 const Cache = require('./cache');
 const spidTalk = require('./spidTalk');
 
-const DEFAULT_CACHE_EXPIRES_IN = 0;
+const DEFAULT_CACHE_EXPIRES_IN = 30;
 const globalWindow = window;
 
 /**
@@ -59,13 +59,13 @@ class Monetization extends EventEmitter {
      * @returns {Object} The data object returned from Schibsted Account
      */
     async hasProduct(productId, spId) {
-        const cacheKey = `prd_${productId}`;
+        const cacheKey = `prd_${productId}_${spId}`;
         const cachedVal = this.cache.get(cacheKey);
         if (cachedVal) {
             return cachedVal;
         }
         const params = { product_id: productId }
-        if (spId !== undefined) {
+        if (spId) {
             params.sp_id = spId;
         }
         const data = await this._spid.get('ajax/hasproduct.js', params);
@@ -85,13 +85,13 @@ class Monetization extends EventEmitter {
      * @returns {Object} The data object returned from Schibsted Account
      */
     async hasSubscription(subscriptionId, spId) {
-        const cacheKey = `sub_${subscriptionId}`;
+        const cacheKey = `sub_${subscriptionId}_${spId}`;
         const cachedVal = this.cache.get(cacheKey);
         if (cachedVal) {
             return cachedVal;
         }
         const params = { product_id: subscriptionId }
-        if (spId !== undefined) {
+        if (spId) {
             params.sp_id = spId;
         }
         const data = await this._spid.get('ajax/hassubscription.js', params);